Split tests into separate files out of mono test file
